INTERLOK-3424 Make EvenOddCondition public
